move functions into internal (#333)
* move functions into internal

* fix
