+++ internal/unmarshal.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"github.com/apache/arrow/go/v16/arrow"
 	"github.com/apache/arrow/go/v16/arrow/array"
+	"github.com/chalk-ai/chalk-go/internal/colls"
 	"github.com/chalk-ai/chalk-go/internal/ptr"
 	"github.com/cockroachdb/errors"
 	"os"
@@ -43,6 +44,8 @@ type NamespaceMemo struct {
 type AllNamespaceMemoT sync.Map
 
 var AllNamespaceMemo = &AllNamespaceMemoT{}
+
+var FieldNotFoundError = errors.New("field not found")
 
 func init() {
 	if chunkSizeStr := os.Getenv(tableReaderChunkSizeKey); chunkSizeStr != "" {
@@ -664,6 +667,511 @@ func SetMapEntryValue(mapValue reflect.Value, key string, value any, allMemo *Al
 		return errors.Wrap(err, "error getting reflect value for map entry")
 	}
 	mapValue.SetMapIndex(reflect.ValueOf(key), ReflectPtr(*rVal))
+	return nil
+}
+
+func ConvertIfHasManyMap(value any) (any, error) {
+	// For has-many values, we get this back:
+	//
+	// {
+	//   "columns": ["user.id", "user.email"],
+	//   "values": [
+	//     ["id1", "id2"],
+	//     ["email1@geemail.com", "email2@geemail.com"]
+	//   ]
+	// }
+	//
+	// We want to convert this to:
+	//
+	// [
+	//   {"user.id": "id1", "user.email": "email1@geemail.com"},
+	//   {"user.id": "id2", "user.email": "email2@geemail.com"}
+	// ]
+	//
+	hasMany, ok := value.(map[string]any)
+	if !ok {
+		return value, nil
+	}
+
+	columnsRaw, hasColumns := hasMany["columns"]
+	valuesRaw, hasValues := hasMany["values"]
+	if !hasColumns || !hasValues {
+		return value, nil
+	}
+
+	columnsAny, ok := columnsRaw.([]any)
+	if !ok {
+		return nil, errors.New("failed to convert columns to []any")
+	}
+
+	columns := make([]string, len(columnsAny))
+	for i, column := range columnsAny {
+		columns[i], ok = column.(string)
+		if !ok {
+			return nil, errors.Newf("failed to convert column '%v' to string", column)
+		}
+	}
+
+	valuesAny, ok := valuesRaw.([]any)
+	if !ok {
+		return nil, errors.New("failed to convert values to [][]any")
+	}
+
+	values := make([][]any, len(valuesAny))
+	for i, row := range valuesAny {
+		values[i], ok = row.([]any)
+		if !ok {
+			return nil, errors.Newf("failed to convert row '%v' to []any", row)
+		}
+	}
+
+	if len(values) == 0 {
+		return nil, errors.New("values of has-many results is empty")
+	}
+	numRows := len(values[0])
+
+	newValues := make([]map[string]any, numRows)
+	for rowIdx := 0; rowIdx < numRows; rowIdx++ {
+		newRow := make(map[string]any)
+		for colIdx, colName := range columns {
+			newRow[colName] = values[colIdx][rowIdx]
+		}
+		newValues[rowIdx] = newRow
+	}
+	return newValues, nil
+}
+
+type InitScope struct {
+	Children map[string]*InitScope
+}
+
+func (s *InitScope) addStr(fqn string) {
+	s.add(strings.Split(fqn, "."))
+}
+
+func (s *InitScope) add(fqnParts []string) {
+	if len(fqnParts) == 0 {
+		return
+	}
+	firstPart := fqnParts[0]
+	if s.Children == nil {
+		s.Children = map[string]*InitScope{}
+	}
+	if _, found := s.Children[firstPart]; !found {
+		s.Children[firstPart] = &InitScope{}
+	}
+	s.Children[firstPart].add(fqnParts[1:])
+}
+
+func BuildScope(fqns []string) (*InitScope, error) {
+	root := &InitScope{}
+	for _, fqn := range fqns {
+		root.addStr(fqn)
+	}
+	return root, nil
+}
+
+func InitRemoteFeatureMap(
+	remoteFeatureMap map[string][]reflect.Value,
+	structValue reflect.Value,
+	cumulativeFqn string,
+	visited map[string]bool,
+	scope *InitScope,
+	allMemo *AllNamespaceMemoT,
+	scopeToJustStructs bool,
+) error {
+	if structValue.Kind() != reflect.Struct {
+		return fmt.Errorf(
+			"feature initialization function argument must be a reflect.Value"+
+				" of the kind reflect.Struct, found %s instead",
+			structValue.Kind().String(),
+		)
+	}
+
+	structName := structValue.Type().Name()
+	if isVisited, ok := visited[structName]; ok && isVisited {
+		// Found a cycle. Just return.
+		return nil
+	}
+	visited[structName] = true
+	defer func() {
+		visited[structName] = false
+	}()
+
+	memo, ok := allMemo.Load(structValue.Type())
+	if !ok {
+		return fmt.Errorf("could not find memo for struct %s, found keys: %v", structName, allMemo.Keys())
+	}
+
+	var fieldNames []string
+	if scopeToJustStructs {
+		fieldNames = colls.Keys(memo.StructFieldsSet)
+	} else {
+		fieldNames = colls.Keys(scope.Children)
+	}
+
+	for _, resolvedFieldName := range fieldNames {
+		nextScope, inScope := scope.Children[resolvedFieldName]
+		if !inScope {
+			continue
+		}
+		updatedFqn := cumulativeFqn + "." + resolvedFieldName
+		fieldIndices, ok := memo.ResolvedFieldNameToIndices[resolvedFieldName]
+		if !ok {
+			// We arrive here when chalk-go receives a response that contains a feature
+			// newly added to one of their has-one feature classes. They have not updated
+			// their codegen'd structs yet, so we simply skip unmarshalling this new
+			// feature to ensure forward compatibility.
+			continue
+		}
+
+		for _, fieldIdx := range fieldIndices {
+			f := structValue.Field(fieldIdx)
+
+			if _, isStruct := memo.StructFieldsSet[resolvedFieldName]; isStruct {
+				if !f.CanSet() {
+					continue
+				}
+				if f.IsNil() {
+					featureSet := reflect.New(f.Type().Elem())
+					f.Set(featureSet)
+				}
+				if err := InitRemoteFeatureMap(
+					remoteFeatureMap,
+					f.Elem(),
+					updatedFqn,
+					visited,
+					nextScope,
+					allMemo,
+					false,
+				); err != nil {
+					return err
+				}
+			} else {
+				remoteFeatureMap[updatedFqn] = append(remoteFeatureMap[updatedFqn], f)
+			}
+		}
+	}
+	return nil
+}
+
+func setFeatureSingle(field reflect.Value, fqn string, value any, allMemo *AllNamespaceMemoT) error {
+	if field.Type().Kind() == reflect.Ptr {
+		rVal, err := GetReflectValue(&value, field.Type(), allMemo)
+		if err != nil {
+			return errors.Wrapf(err, "getting reflect value for feature '%s'", fqn)
+		}
+		field.Set(*rVal)
+		return nil
+	} else if field.Kind() == reflect.Map {
+		bucket, err := GetBucketFromFqn(fqn)
+		if err != nil {
+			return errors.Wrapf(err, "extracting bucket value for feature '%s'", fqn)
+		}
+		if err := SetMapEntryValue(field, bucket, value, allMemo); err != nil {
+			return errors.Wrapf(err, "setting map entry value for feature '%s'", fqn)
+		}
+		return nil
+	} else {
+		return fmt.Errorf("expected a pointer type for feature '%s', found %s", fqn, field.Type().Kind())
+	}
+}
+
+// ThinUnmarshalInto is called per row. Any operation that can be
+// done outside of this function must be done outside of this function.
+func ThinUnmarshalInto(
+	structValue reflect.Value,
+	fqnToValue map[string]any,
+	namespace string,
+	namespaceScope *InitScope,
+	namespaceMemo *NamespaceMemo,
+	allMemo *AllNamespaceMemoT,
+) (returnErr error) {
+	remoteFeatureMap := map[string][]reflect.Value{}
+	if err := InitRemoteFeatureMap(
+		remoteFeatureMap,
+		structValue,
+		namespace,
+		map[string]bool{},
+		namespaceScope,
+		allMemo,
+		true,
+	); err != nil {
+		return errors.Wrap(err, "initializing result holder struct")
+	}
+
+	for fqn, value := range fqnToValue {
+		targetFields, ok := remoteFeatureMap[fqn]
+		if !ok {
+			// If not a has-one remote feature, e.g. user.account.balance
+			fieldIndices, ok := namespaceMemo.ResolvedFieldNameToIndices[fqn]
+			if !ok {
+				// For forward compatibility, i.e. when clients add
+				// more fields to their dataclasses in chalkpy, we want
+				// to default to not erring when trying to deserialize
+				// a new field that does not yet exist in the Go struct.
+				// Eventually we might consider exposing a flag.
+				continue
+			}
+			for _, fieldIdx := range fieldIndices {
+				targetFields = append(targetFields, structValue.Field(fieldIdx))
+			}
+		}
+
+		for _, field := range targetFields {
+			if value == nil {
+				if field.Type().Kind() == reflect.Map && field.IsNil() {
+					field.Set(reflect.MakeMap(field.Type()))
+					continue
+				}
+
+				// TODO: Add validation for optional fields
+				continue
+			}
+			if err := setFeatureSingle(field, fqn, value, allMemo); err != nil {
+				structName := structValue.Type().String()
+				outputNamespace := "unknown namespace"
+				sections := strings.Split(fqn, ".")
+				if len(sections) > 0 {
+					outputNamespace = sections[0]
+				}
+				if errors.Is(err, FieldNotFoundError) {
+					fieldError := fmt.Sprintf("Error unmarshaling feature '%s' into the struct '%s'. ", fqn, structName)
+					fieldError += fmt.Sprintf("First, check if you are passing a pointer to a struct that represents the output namespace '%s'. ", outputNamespace)
+					fieldError += fmt.Sprintf("Also, make sure the feature name can be traced to a field in the struct '%s' and or its nested structs.", structName)
+					return errors.New(fieldError)
+				} else {
+					return errors.Wrapf(err, "unmarshaling feature '%s' into the struct '%s'", fqn, structName)
+				}
+			}
+		}
+
+	}
+	return nil
+}
+
+type UnmarshalChunkResult struct {
+	chunkIdx int
+	rows     []reflect.Value
+	err      error
+}
+
+func UnmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) {
+	defer func() {
+		if panicContents := recover(); panicContents != nil {
+			detail := "details irretrievable"
+			switch typedContents := panicContents.(type) {
+			case *reflect.ValueError:
+				detail = typedContents.Error()
+			case string:
+				detail = typedContents
+			}
+			returnErr = errors.Newf("exception occurred while unmarshalling result: %s", detail)
+		}
+	}()
+
+	numRows, err := Int64ToInt(table.NumRows())
+	if err != nil {
+		return errors.Newf("table too large to unmarshal, found %d rows", table.NumRows())
+	}
+
+	slicePtr := reflect.ValueOf(resultHolders)
+	if slicePtr.Kind() != reflect.Ptr {
+		return fmt.Errorf(
+			"result holder should be a pointer to a slice of structs, "+
+				"got '%s' instead",
+			slicePtr.Kind(),
+		)
+	}
+
+	slice := reflect.Indirect(slicePtr)
+	if slice.Kind() != reflect.Slice {
+		return fmt.Errorf(
+			"result holder should be a pointer to a slice of structs, "+
+				"got '%s' instead",
+			slice.Kind(),
+		)
+	}
+
+	sliceElemType := slice.Type().Elem()
+	if sliceElemType.Kind() != reflect.Struct {
+		return fmt.Errorf(
+			"result holder should be a pointer to a slice of structs, "+
+				"got a pointer to a slice of '%s' instead",
+			sliceElemType.Kind(),
+		)
+	}
+
+	rows, _, scalarsErr := ExtractFeaturesFromTable(table, false)
+	if scalarsErr != nil {
+		return scalarsErr
+	}
+	if len(rows) == 0 {
+		return nil
+	}
+
+	scope, err := BuildScope(colls.Keys(rows[0]))
+	if err != nil {
+		return errors.Wrap(err, "building deserialization scope")
+	}
+
+	allMemo := AllNamespaceMemo
+	if err := PopulateAllNamespaceMemo(sliceElemType, nil); err != nil {
+		return errors.Wrap(err, "building namespace memo")
+	}
+
+	structName := sliceElemType.Name()
+	namespace := ChalkpySnakeCase(structName)
+	nsScope := scope.Children[namespace]
+
+	var rowToStruct func(map[string]any) (*reflect.Value, error)
+	if nsScope != nil {
+		// single namespace unmarshalling
+		nsMemo, ok := allMemo.Load(sliceElemType)
+		if !ok {
+			return errors.Newf("namespace '%s' not found in memo, found keys: %v", structName, allMemo.Keys())
+		}
+
+		rowToStruct = func(row map[string]any) (*reflect.Value, error) {
+			featuresStruct := reflect.New(sliceElemType)
+			if err := ThinUnmarshalInto(
+				featuresStruct.Elem(),
+				row,
+				namespace,
+				nsScope,
+				nsMemo,
+				allMemo,
+			); err != nil {
+				return nil, err
+			}
+			return ptr.Ptr(featuresStruct.Elem()), nil
+		}
+	} else {
+		// Multi namespace unmarshalling
+		type namespaceMetaT struct {
+			fieldIdx  int
+			namespace string
+			scope     *InitScope
+			memo      *NamespaceMemo
+		}
+
+		namespaceMeta := []namespaceMetaT{}
+		for i := 0; i < sliceElemType.NumField(); i++ {
+			fieldMeta := sliceElemType.Field(i)
+			if fieldMeta.Type.Kind() != reflect.Struct {
+				return errors.Newf(
+					"If attempting single namespace unmarshalling, please make sure you're unmarshalling into the correct struct. "+
+						"Attempted single namespace unmarshalling into struct '%s', but results are from these namespaces: %v. "+
+						"If attempting multi-namespace unmarshalling, please pass in a pointer to a struct whose fields are all "+
+						"structs (not struct pointers) corresponding to the output namespaces. The problematic field is '%s' of type '%s'.",
+					structName,
+					colls.Keys(scope.Children),
+					fieldMeta.Name,
+					fieldMeta.Type.Name(),
+				)
+			}
+
+			fieldNamespace := ChalkpySnakeCase(fieldMeta.Type.Name())
+
+			fieldScope := scope.Children[fieldNamespace]
+			if fieldScope == nil {
+				return errors.Newf(
+					"Please make sure you're unmarshalling into the correct struct. Attempted single namespace "+
+						"unmarshalling into struct '%s', and attempted multi-namespace unmarshalling into the field '%s' "+
+						"of type '%s', but results are from these namespaces: %v",
+					structName,
+					fieldMeta.Name,
+					fieldMeta.Type.Name(),
+					colls.Keys(scope.Children),
+				)
+			}
+
+			fieldMemo, ok := allMemo.Load(fieldMeta.Type)
+			if !ok {
+				return errors.Newf("namespace '%s' not found in memo, found keys: %v", structName, allMemo.Keys())
+			}
+
+			namespaceMeta = append(namespaceMeta, namespaceMetaT{
+				fieldIdx:  i,
+				namespace: fieldNamespace,
+				scope:     fieldScope,
+				memo:      fieldMemo,
+			})
+		}
+
+		rowToStruct = func(row map[string]any) (*reflect.Value, error) {
+			rootStruct := reflect.New(sliceElemType)
+			for _, meta := range namespaceMeta {
+				if err := ThinUnmarshalInto(
+					rootStruct.Elem().Field(meta.fieldIdx),
+					row,
+					meta.namespace,
+					meta.scope,
+					meta.memo,
+					allMemo,
+				); err != nil {
+					return nil, errors.Wrapf(
+						err,
+						"error unmarshalling into field '%s'",
+						sliceElemType.Field(meta.fieldIdx).Name,
+					)
+				}
+			}
+			return ptr.Ptr(rootStruct.Elem()), nil
+		}
+	}
+
+	var wg sync.WaitGroup
+	numWorkers := runtime.NumCPU()
+	resChan := make(chan *UnmarshalChunkResult, numWorkers)
+	chunkSize := (len(rows) / numWorkers) + 1
+
+	for chunkIdx := 0; (chunkIdx * chunkSize) < len(rows); chunkIdx += 1 {
+		wg.Add(1)
+		go func(routineChunkIdx int) {
+			defer wg.Done()
+			chunkStart := routineChunkIdx * chunkSize
+			chunkEnd := chunkStart + chunkSize
+			chunkRows := rows[chunkStart:min(chunkEnd, len(rows))]
+			results := make([]reflect.Value, len(chunkRows))
+			for rowIdx, row := range chunkRows {
+				res, err := rowToStruct(row)
+				if err != nil {
+					resChan <- &UnmarshalChunkResult{chunkIdx: routineChunkIdx, err: err}
+					return
+				} else {
+					results[rowIdx] = *res
+				}
+			}
+			resChan <- &UnmarshalChunkResult{chunkIdx: routineChunkIdx, rows: results}
+		}(chunkIdx)
+	}
+
+	wg.Wait()
+	close(resChan)
+
+	var allChunks []*UnmarshalChunkResult
+	for chunkResult := range resChan {
+		allChunks = append(allChunks, chunkResult)
+	}
+
+	sort.Slice(allChunks, func(i, j int) bool {
+		return allChunks[i].chunkIdx < allChunks[j].chunkIdx
+	})
+
+	newSlice := reflect.MakeSlice(slice.Type(), numRows, numRows)
+	rowIdx := 0
+	for _, chunkResult := range allChunks {
+		if chunkResult.err != nil {
+			return chunkResult.err
+		}
+		for _, row := range chunkResult.rows {
+			newSlice.Index(rowIdx).Set(row)
+			rowIdx += 1
+		}
+	}
+	slice.Set(newSlice)
+
 	return nil
 }
